pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-ads-mcp-server"
version = "1.0.0"
description = "MCP Server para consultar datos de Meta Ads (Facebook/Instagram) desde Claude"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Tu Equipo", email = "team@tuempresa.com" }
]
license = { text = "MIT" }
dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[project.scripts]
meta-mcp-server = "meta_ads_mcp:main"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]